Import up to date RBTools. Fix exporter to export Tests
This pulls in many files from our current project. Since we want tests in those projects, it also merged in tests with the rest of the RBTools. In addition, I created an option to export the tool without the tests for public packages. But we can continue to export the tool with tests for internal projects.